Refactor register access to use long address type
Changed ReadRegister and WriteRegister methods across all emulator classes and interfaces to use 'long' instead of 'uint' for address parameters, improving support for 64-bit address spaces. Updated related struct fields and memory size properties to 'long' where appropriate. Added HardwareHealthProbe utility and enhanced UI with drop shadow and transparency. InstructionTranslator class expanded with stub handlers for ARM/MIPS instructions.

+++ ComcastX1Emulator.cs
@@ -280,13 +280,13 @@ namespace ProcessorEmulator
             return Initialize().Result;
         }
 
-        public byte[] ReadRegister(uint address)
+    public byte[] ReadRegister(long address)
         {
             // Universal register access - works with any architecture
             return new byte[8]; // Return 64-bit register value
         }
 
-        public void WriteRegister(uint address, byte[] data)
+    public void WriteRegister(long address, byte[] data)
         {
             // Universal register writing - works with any architecture
         }
